vgg16/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bagpy==0.5
bitstring==4.0.1
blinker==1.4
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
comm==0.1.3
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
flatbuffers==2.0
flit_core==3.8.0
fonttools==4.25.0
frozenlist==1.3.3
gast==0.4.0
gnupg==2.3.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.7.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.0.3
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
Owlready2==0.37
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip==23.0.1
platformdirs==3.2.0
ply==3.11
pooch==1.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrealsense2==2.53.1.4623
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
setuptools-scm==7.1.0
sip==6.6.2
six==1.16.0
stack-data==0.6.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
termcolor==2.1.0
toml==0.10.2
tomli==2.0.1
torch==1.12.0
torch-tb-profiler==0.4.1
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
yarl==1.8.1
zipp==3.15.0
